+++ src/Phel.php
@@ -6,22 +6,23 @@ use Phel\Lang\Collections\HashSet\PersistentHashSetInterface;
 use Phel\Lang\Collections\LinkedList\PersistentListInterface;
 use Phel\Lang\Collections\Map\PersistentMapInterface;
 use Phel\Lang\Collections\Vector\PersistentVectorInterface;
+use Phel\Lang\Keyword;
 use Phel\Lang\Registry;
+use Phel\Lang\Symbol;
 use Phel\Lang\TypeFactory;
+use Phel\Lang\Variable;
 use Phel\Phel as InternalPhel;
 
 /**
  * Public API for Phel.
  *
  * @mixin Registry
- * @mixin TypeFactory
  */
 final class Phel extends InternalPhel
 {
     /**
      * Proxy undefined static method calls to
      * - {@see Registry} singleton.
-     * - {@see TypeFactory} singleton.
      *
      * @param  string  $name
      * @param  list<mixed>  $arguments
@@ -35,12 +36,21 @@ final class Phel extends InternalPhel
             return $registry->$name(...$arguments);
         }
 
-        $typeFactory = TypeFactory::getInstance();
-        if (is_callable([$typeFactory, $name])) {
-            return $typeFactory->$name(...$arguments);
-        }
-
         throw new BadMethodCallException(sprintf('Method "%s" does not exist', $name));
+    }
+
+    /**
+     * Get a reference to a stored definition. This is part of the Registry but has to be redefined
+     * because it is returning the reference to the definition.
+     *
+     * @noinspection PhpUnused
+     * @see GlobalVarEmitter
+     *
+     * @return mixed Reference to the stored definition.
+     */
+    public static function &getDefinitionReference(string $ns, string $name): mixed
+    {
+        return Registry::getInstance()->getDefinitionReference($ns, $name);
     }
 
     /**
@@ -89,15 +99,22 @@ final class Phel extends InternalPhel
     }
 
     /**
-     * Get a reference to a stored definition.
-     *
-     * @noinspection PhpUnused
-     * @see GlobalVarEmitter
-     *
-     * @return mixed Reference to the stored definition.
+     * @template T
+     * @param T $value The initial value of the variable
+     * @return Variable<T>
      */
-    public static function &getDefinitionReference(string $ns, string $name): mixed
+    public static function variable($value, ?PersistentMapInterface $meta = null): Variable
     {
-        return Registry::getInstance()->getDefinitionReference($ns, $name);
+        return new Variable($meta, $value);
+    }
+
+    public static function symbol(string $name): Symbol
+    {
+        return Symbol::create($name);
+    }
+
+    public static function keyword(string $name, ?string $namespace = null): Keyword
+    {
+        return Keyword::create($name, $namespace);
     }
 }