+++ src/php/Lang/TypeFactory.php
@@ -76,28 +76,6 @@ final class TypeFactory
         return PersistentVector::fromArray($this->hasher, $this->equalizer, $values);
     }
 
-    /**
-     * @template T
-     *
-     * @param T $value The initial value of the variable
-     *
-     * @return Variable<T>
-     */
-    public function variable($value): Variable
-    {
-        return new Variable(null, $value);
-    }
-
-    public function symbol(string $name): Symbol
-    {
-        return Symbol::create($name);
-    }
-
-    public function keyword(string $name, ?string $namespace = null): Keyword
-    {
-        return Keyword::create($name, $namespace);
-    }
-
     public function getEqualizer(): EqualizerInterface
     {
         return $this->equalizer;